% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/letter.R
\name{letter}
\alias{letter}
\title{generating a letter}
\usage{
letter(x)
}
\arguments{
\item{x}{A useless parameter for the parallel computition}
}
\value{
a random letter of the alphabet
}
\description{
sample from the alphabet
}
